Mar 4, 2023, 1:12 PM

+++ lib/model/data/monitoring_state.dart
@@ -4,7 +4,7 @@ part 'monitoring_state.freezed.dart';
 
 @freezed
 class MonitoringState with _$MonitoringState {
-  const factory MonitoringState.success(double battery, [String? msg]) =
+  const factory MonitoringState.success(int battery, [String? msg]) =
       Success;
   const factory MonitoringState.fail(String msg) = Fail;
 }
